+++ pokerole_irl/user_app/urls.py
@@ -4,5 +4,6 @@ from .views import RegisterView, UserProfileView
 
 urlpatterns = [
     path('register/', RegisterView.as_view(), name='register'),
+    path('profile/<str:slug>', UserProfileView.as_view(), name='user-profile'),
     path('profile/', UserProfileView.as_view(), name='user-profile')
 ]