+++ pokerole_irl/worlds_app/models.py
@@ -1,27 +1,45 @@
 from django.db import models
+from django.db.models.constraints import UniqueConstraint
 from django.contrib.auth.models import User
+from django.template.defaultfilters import slugify
 
-# Create your models here.
+
 class World(models.Model):
     members = models.ManyToManyField(User, through='WorldMember')
     name = models.CharField(max_length=100)
     description = models.TextField(default="")
-
+    slug = models.SlugField(blank=True)
 
     def __str__(self):
         return self.name
 
+    def save(self, **kwargs):
+        self.slug = slugify(self.name)
+        super().save(**kwargs)
+
+    UniqueConstraint(fields=["name"], name="name_constraint")
+
+
 class Character(models.Model):
     owner = models.ForeignKey(User, on_delete=models.CASCADE)
-    world = models.ForeignKey(World, blank=True, null=True, on_delete=models.SET_NULL)
+    world = models.ForeignKey(
+        World, blank=True, null=True, on_delete=models.SET_NULL)
     first_name = models.CharField(max_length=100)
     last_name = models.CharField(max_length=100)
     description = models.TextField(default="")
-
+    slug = models.SlugField(blank=True)
 
     def __str__(self):
         return f"{self.first_name} {self.last_name}"
-    
+
+    def save(self, **kwargs):
+        self.slug = slugify(str(self))
+        super().save(**kwargs)
+
+    UniqueConstraint(fields=["world", "first_name",
+                     "last_name"], name="name_world_constraint")
+
+
 class WorldMember(models.Model):
     user = models.ForeignKey(User, null=True, on_delete=models.SET_NULL)
     world = models.ForeignKey(World, on_delete=models.CASCADE)
@@ -29,6 +47,6 @@ class WorldMember(models.Model):
         "ROLE", "Owner Member")
     role = models.CharField(
         max_length=30, choices=ROLES.choices, default="Member")
-    
+
     def __str__(self):
         return f"{self.user}"